Update patch set 14
Patch Set 14: (1 inline comment)

Whether restoring the lstrip() or not would make a difference. Functionaly probably not, but logically I would consider it bad practice to change behaviour in an unassociated commit, since it becomes more difficult for people looking back at the git log to easily understand the intent of a code change.

Patch-set: 14