Update patch set 1
Patch Set 1: (3 inline comments)

I think you mis-understood, the lstrip() was not added to work around a minidom issue, it was added when the minidom issue was fixed and the broken regex was dropped to ensure the following:

If user set a description and the magic string was being added, the user description and custom magic string would be separated and clearly distinguishable. (optional)

If the user did not enter a custom description, so only the magic string was used, the resulting description would not be surrounded by excess whitespace exactly as if you used jenkins' UI to set the description (hard requiredment)


I'm not wedded to holding onto the lstrip(), but if you want to remove it, then you should remove the leading '\n\n' from the magic managed string and have no easily visible distinction between the user specified part and the generated part.

The simple hard requirement is that there should be no leading whitespace in the final string between the description tags in the XML.

Patch-set: 1