Update patch set 5
Patch Set 5: I would prefer that you didn't merge this

I tried testing this but unfortunately multi scm doesn't work in jenkins.  The xml gets generated correctly but I think jenkins has a problem reading multiple scm entries.  In fact it's not possible to setup multiple git scm from the jenkins UI.   I'm using jenkins ver 1.5.3.2 (LTS) and git plugin ver 2.0

So although this change seems like an improvement because the generated xml is correct it is misleading because this feature doesn't work in jenkins.

here is my test config.  After creating this job I view the job config in the Jenkins UI but the SCM is set to 'None'.     

- job:
    name: test-multi-scm

    scm:
      - git:
          url: 'https://git.openstack.org/openstack-infra/jeepyb'
          basedir: 'jeepyb'
      - git:
          url: 'https://git.openstack.org/openstack-infra/gerritlib'
          basedir: 'gerritlib'

Patch-set: 5