Update patch set 1
Patch Set 1: I would prefer that you didn't merge this

I think this should have a note somewhere indicating the zuul parameters need only be explicitly listed if you are not using zuul with gearman. When used with gearman these are not required (but may be nice to have).

Patch-set: 1
Label: Code-Review=-1