Make email-ext's AlwaysTrigger the default.
This matches the behavior of the email-ext plugin in Jenkins as of
version 2.36.

There is some discussion on the original issue about whether this is a
desirable default, however, so email-ext may change in the future.  See
https://issues.jenkins-ci.org/browse/JENKINS-20013 for more
information.

Update: Convert tab to space per review comment.  Also corrected the
same problem in email-ext001.xml where it originated.

Change-Id: Id6dfb5fe5930cea3efc88870d9747197717c5b19