Update patch set 1
Patch Set 1:

Well, this patch only parallelizes the creation/updating of the jobs, so no problems tehre, it also makes it easy to parallelize any other function.

BUT, it uses threads and not processes, so not real multiprocessing on standard cpython vm. But the issue is with io (when calling jenkins) so it improves performance dramatically anyhow, at least for us where the issue is the call complete time.

If you want, I can try to migrate it to multiprocess instead.

Patch-set: 1
Label: Code-Review=0