Update patch set 2
Patch Set 2: I would prefer that you didn't merge this

would it be possible to add a test for lists and dicts?  also it would be great if you can make a literalinclude reference to the test files in the jjb documentation because I don't think users would know how to do this otherwise.

Patch-set: 2
Label: Code-Review=-1
Label: Workflow=0