Update patch set 3
Patch Set 3: I would prefer that you didn't merge this

This is a weak -1.  I agree that you could keep using this to make triggering via the web UI easier, but Zuul does have a 'trigger-job' script that can do the same thing, and it will be easier to keep it in sync with Zuul in the future.

So I'd be in favor of just removing the module from JJB, but if you feel strongly about keeping it, that's okay too.

Patch-set: 3
Label: Code-Review=-1