Update patch set 1
Patch Set 1: I would prefer that you didn't merge this

(8 inline comments)

Glad to see someone else has picked up the need for threads to improve performance on large systems.

Main concern is the lack of tests to be run automatically, which was the biggest thing missing from my attempt as well. I think it's important that there is some way to test both single threaded (should be the default behaviour) as well as multi-threaded to try and catch any concurrency problems that may be accidentally added. Probably worth looking at how to solve this before considering whether to use the simplier interface for threadpool management from the multiprocessing module.

Note that the multiprocessing module in python as well as providing an API for parallel processing using separate processes, also provides a similar API for parallel processing using threads.

Patch-set: 1
Label: Code-Review=-1