Add support for URLTrigger plugin
The URLTrigger plugin has fairly complex configuration options, so
this code is somewhat involved.  I tried to keep it simple and
understandable without introducing too much duplication.

It is also the first significant amount of Python I've ever written,
so any feedback on how to clean this up and make it more idiomatic is
more than welcome.  I extracted more methods than the other components
in this project typically use; I'm not sure if that's OK here.

Update 2014-06-24:
  * Rebase on latest master
  * Use a lookup table rather than dynamic method lookup, per
    reviewer comments.

Update 2014-06-26:
 * Modify formatting per reviewer comment

Change-Id: Ibc96b3ecd65afb27398328f2b6220c32a800f3e6