Job name override - used when storing
This allows a job to specify the name that it should use on
the Jenkins system, if different from the 'name' specified
in the job definition.

This allows you to have e.g. an "old" version of a job at
the same time that the "new" version is present, and gradually
move projects from the "old" job to the "new" one, without having
to rename the jobs on the Jenkins machine and update references
to the job in other (e.g upstream) jobs.

Change-Id: I793a046e0206053e126bbfbc622876ded5267135