Update patch set 1
Patch Set 1:

The idea here is to let the core routines accept a file object as well as a filename.

The motivation for this is that we are building a tool on top of jenkins-job-builder and our tool generates YAML and then needs to pass it to Builder.update_job. Currently it does this by writing a temp file and then passing it the filename. But it would be a little nicer if we could avoid the temp file and simply pass in a StringIO object.

Patch-set: 1
Label: Code-Review=0