Update patch set 1
Patch Set 1:

Apart from the tests, I've been trying to replace the threading module by the multiprocessing module and I'm having lots of issues trying to keep the decorator approach.
It seems that decorated functions are not so easily pickable, If anyone has experience with it I'd be more than glad to hear it :)
In the meantime I'll send the threading based one with tests, and maybe leave the switch to multiprocessing later, for most of the cases the downside is not that big, as the main time consuming tasks are the io related ones and those run in parallel also for threads.

Patch-set: 1