Update patch set 1
Patch Set 1:

Marc wrote:
> The addition of a .travis.yml lets me push my changes to my own fork and then run Travis CI on them. This gives me an additional level of confidence that my stuff works before I push it to Gerrit.

Hi! What I do is to simply run tox before submitting, that catch almost all the issue and save up time.  Whenever a patch is known to break tests but I need some others to review it, I flag it as a work in progress.

Patch-set: 1