Update patch set 4
Patch Set 4: I would prefer that you didn't merge this

(1 inline comment)

Peter Jönsson wrote:
> What remains to get this merged? Thanks!

Each proposed patch has to be validated by two different core reviewers, more or less the equivalent of having merge right.  That usually ensure higher code quality and a lot of review :-]


The error message duplicate the list of keys being validated. You can display the dict keys in the error message, this way later on we will only have to add the new SCM in two places (in the doc and in the dict).   That is very minor but nice to have.

Patch-set: 4
Label: Code-Review=-1
Label: Workflow=0