Allow to specify multiple branches per project in Gerrit trigger
Allow to specify multiple branch patterns in a 'branches' block
within the 'project' block.

If no 'branches' block is present, fall back to using the single
branch specified directly in the 'project' section.

If a branch is specified both at the project level, and in the
'branches' section, ignore the one at project level.

Bug: #1302502
Change-Id: I1fb2ce0bb3776d365f2c0e98a703c6dfba5fac7f