Update patch set 3
Patch Set 3: Looks good to me, but someone else must approve

(1 inline comment)

Have to admit I didn't understand the purpose behind it at first glance, but the additional of the documentation helped me see that it would be very useful for condensing a large number of similar templates into a set much easier managed.

Added a note indicating that while it might be nice to use the tests as documentation, I don't think it's worth reworking how the get_scenarios test code right now to save a few lines of text in this case.

Patch-set: 3