Support excludes when recursively processing paths
Include an additional option to exclude some paths/patterns from being
included in the search to allow a complex hierarchy of directories to be
referenced under a single path. Provide support to allow the user to
specify patterns, relative and/or absolute path references.

Change load_files function to process list of files or directories.

Change-Id: I236b364c268fd2bf5511a2c6d9a1c87914f3b086