Update patch set 7
Patch Set 7: I would prefer that you didn't merge this

(7 inline comments)

I love the idea.I don't think we need to support '-' for output/path since they are the default for output and path parameters.  I gave you a bunch of suggestions to simplify the code.

Found out a bug while testing which is that the test command will output jobs to stdout when we pass it job names regardless of the -o parameters.  Although it is a pre existent bug, you might want to fix it in this patch as well :]

+1 on the idea \O/

Patch-set: 7
Label: Code-Review=-1
Label: Workflow=0