Update patch set 4
Patch Set 4: (3 inline comments)

I would recommend to use flake8 instead of pep8 and have its options listed in a [flake8] section of tox.ini (eventually duplicated to a [pep8] section).   flake8 is a wrapper around both pep8 and pyflakes (and mccabe if it is around).

The code change to the publisher should be done in another change I believe.


but yeah, lets get rid of pep8 1.3.3 !

Patch-set: 4
Label: Code-Review=0
Label: Workflow=0