Update patch set 4
Patch Set 4: I would prefer that you didn't merge this

(4 inline comments)

Not sure we should set the plugin='exclusion@0.8' attribute, see inline. Might want to talk with a few more people about it.  Not a blocker anyway.

The links to Jenkins ci are not rendering properly in the doc. For each occurrences I added an URL pointing to the doc generated for this patchset.  Se e inline comment for details :]

Otherwise looks fine.

Patch-set: 4
Label: Code-Review=-1
Label: Workflow=0