Update patch set 2
Patch Set 2:

So I think we should get this patch to create a new builder named git2.

Then in another change start working on Darragh proposal to detect plugins versions on the targeted Jenkins. That will need a bit more work than simply supporting the v2 of the git plugin, but I have the feeling it would be a useful feature for other plugins as well.

Patch-set: 2