Update patch set 3
Patch Set 3:

Hi Khai Do, thanks for your input.

We're building a tool on top of JJB that generates a YAML config meant to be digested by JJB. But with the current implementation, we need to pass the YAML file as a filename which requires us to write it to a temp file so JBB can read from it, and then we delete the temp file. A nicer approach would be to pass our in-memory YAML config file object (e.g., StringIO) without requiring us to write the file to disk.

I'm happy to add/update the documentation if necessary, can you give me pointers on where it would be best to do this?

Thanks, Alex

Patch-set: 3