Update patch set 1
Patch Set 1:

Note that nosetests and py.test (with pytest-logcapture plugin) have log capturing where logged messages are buffered and then only displayed when a test fails. This results in clean output when tests pass but helpful logs when test fail. 

If testr doesn't have this feature then this change is a reasonable substitute.

Patch-set: 1