Update patch set 10
Patch Set 10:

This feature could easily be implemented by a 3rd party software if the "list" command gets merged. In fact, get_jobs is already available for such purposes.

I have the same opinion as Clark on that one, it looks like a very specific feature for a specific need. My fear is that more of those specific need features get proposed and merged in the future. I have always see JJB as a low level tool, not a full blown tool to interact with Jenkins over the CLI.

Patch-set: 10