Update patch set 7
Patch Set 7:

I'm still looking at the decorator approach, but it seems at least on the tests I did that it's slower (probably the overhead of creating full processes), and as the bottleneck is io (the connectivity to jenkins) I'm not sure if it's so relevant to use processes instead of threads.

I'll do some more tests.

btw. I like the code getting simpler though

Patch-set: 7