Update patch set 3
Patch Set 3:

Yes, I agree, it's part of the features introduced by that patchset.

I just rebased so it will not fail the jenkins jobs.

I think that it might be faster to apply the reursive and the exclude options separated, but if that will not be an issue, I don't care if this one or Darragh's patch is merged as long as it is :)

Patch-set: 3
Label: Code-Review=0