Update patch set 3
Patch Set 3: I would prefer that you didn't merge this

(2 inline comments)

@Mathieu I think there has been some back and forth on both approaches :) Some functionality from this one has been included in the other

I'd be quite happy for the recursive behaviour to go in via this change and for the other one to just add the 'exclude', as I think that will take more work to get right.

I do think it would be better to try and keep the recursive path processing outside of the Builder class and allow the calling method to control the list of paths to be processed.

Patch-set: 3
Label: Code-Review=-1