Update patch set 2
Patch Set 2: I would prefer that you didn't merge this

(1 inline comment)

I think the root cause as to why in your case you are hitting this on Fedora 20 while I am not is still unknown. I'll try to do some more detailed testing when I get home, but I suspect that something else may be altering how minidom generates XML similar to how PyXML would patch it dynamically at module load time. For now this fixes your problem because in turn this code is patching it back to it's correct state and wins out because it's occurring later.

I would suggest that we dig in deeper to understand better what is happing as it likely means there needs to be a second condition to check when we should patch the writexml() method instead of just comparing the version.

Would not surprise me if this was the same underlying problem that someone else encountered previously when changing the comparision from less than version 2.7 to less than or equal to version 2.7.3.

Patch-set: 2
Reviewer: Gerrit User 1054 <1054@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1