Update patch set 2
Patch Set 2:

Unfortunately PyXML has been unmaintained since 2004. It fixes the problem nicely, but isn't acceptable because of this and the fact that it won't work with python 3.


So pretty much left with finding what's triggering the problem on Fedora 20 (guessing some additional package) or switching to lxml (likely non trivial).

Patch-set: 2