Update patch set 1
Patch Set 1: Work In Progress

I'll change it to default to a single thread just as soon as i able, be a few days though. Presumably it will be possible to configure so that the tests run multiple threads otherwise problems such as cache corruption, for which the change adds locks to prevent won't be caught in the gate job.

Patch-set: 1