Update patch set 2
Patch Set 2:

Forgot to add that I tried switching to lxml before and what constantly tripped me up was not being able to easily encode the HTML entities that occur within the text data blocks so that it was outputted in a transparent manner when generating the XML.

lxml also alters the quotes use in the xml declaration at the top of the file and I found no real solution to that. Whether it will affect the loading of the code by Jenkins I'm not entirely sure. Only minidom seems to generate output consistent with what jenkins is expecting.

Patch-set: 2