Update patch set 2
Patch Set 2: Workflow-1

Yep, I will modify the modules to use literalinclude to the tests that I added.

Re: testing coverage, I agree that it would be awesome to measure coverage so that we can improve it. I did informally here by using tools that I usually use on other projects -- pytest and pytest-cov. Antoine had suggested using more OpenStack-favored tools like pbr and testr (see: https://review.openstack.org/87382). This should be possible to use pbr and testr but there are a couple of issues in these that are blocking it:

- https://review.openstack.org/89205
- https://code.launchpad.net/~chmouel/testrepository/coverage-package-name/+merge/213218

We can try to move those along and then when they're done, we can do coverage using them.

Actually https://review.openstack.org/87382 kind of works around the bug by creating a symlink so I guess we could merge it and then remove the symlink hack when pbr and testr are updated.

Patch-set: 2
Label: Workflow=-1