Update patch set 12
Patch Set 12:

Why isn't argv made a required argument and sys.argv passed from the __main__ invocation?

I feel it's much simpler, doesn't require adding logic related to default value and 3 lines to explain a python gotcha.

Is it to preserve the previous behaviour? And assuming somewhere, someone imported jenkins_jobs and called main() as if it was a library?

Patch-set: 12
Label: Verified=0