Update patch set 1
Patch Set 1:

IMO, 2 changes are confusing and create unneeded load on reviewers as we have to make sure none are conflicting each others depending on the merge order or what not.

I would have preferred for this particular change to implement the example as test right now instead of creating a new change for it, possibly conflicting each others. Thanks.

Patch-set: 1