Update patch set 5
Patch Set 5:

Yeah, I don't know if jenkins-job-builder should do a ton of validation. It doesn't really have much knowledge of Jenkins and all its plug-ins. Jenkins plugins can evolve pretty quickly so it may be hard to keep up with them. Might not be worth it IMHO.

I'd vote for merging this now and revisiting vaildation if it's something that comes up as being useful.

Patch-set: 5