Update patch set 1
Patch Set 1:

We still don't have a super satisfying solution here. setup_requires causes setuptools to install a pbr egg and then when it checks install_requires it doesn't install pbr in the virtualenv. Then if you try to do anything that imports pbr, it fails.

I almost wonder if we should remove `setup_requires=['pbr']` and instead add some logic to setup.py that craps out and tells you to install pbr if it's not available?

Patch-set: 1