Some tweaks to get closer to Python 3 compat
Need to depend on the earliest released version of python-jenkins that
supports python 3. Uses OrderedDict to ensure that anywhere dicts are
iterated over to provide the order of the XML tags to create when
generating XML that the order of the resulting output XML snippet will
be consistent between different runs.

Change-Id: If4b35e2ceee8239379700e22eb79a3eaa04d6f0f