Some tweaks to get closer to Python 3 compat
Convert to use idioms that work for both python 3 and python 2.6+ and
ensure that a suitable version of dependencies is included for python 3
compatibility.

Update python-jenkins to 0.3.3 as the earliest version that supports
python 3 without any known regressions. Add an extra parser check for
missing 'command' due to changes in how argparse works under python 3.

Where contents should be retained, to access the first element of a dict
in both python 2 and 3, use 'next(iter(dict.items()))' as the standard
idiom to replace 'dict.items()[0]' as 'items()' returns an iterator in
python 3 which cannot be indexed. Using 'next(iter(..))' allows for
both lists and iterators to be passed in without unnecessary conversion
of iterators to lists which would be true of 'list(dict.items())[0]'.

Alternatively, where further access to the data is not required, we can
use 'dict.popitem()'.

Change-Id: If4b35e2ceee8239379700e22eb79a3eaa04d6f0f