Update patch set 1
Patch Set 1:

Ideally, I agree we can share common code. There's a couple of differences when used as a build step.

FWIW, the common code is plagued by if statements to accommodate the various consumers of this function. It makes it unreadable from my PoV. That's the reason, I copied the code to make only one thing properly and easy to read/understand.

I'll refactor the code as requested and you'll judge by yourself.

Patch-set: 1
Label: Verified=0