Update patch set 14
Patch Set 14: Code-Review+1

That is still going to break my integration tests which run on a Ubuntu Precise which has an even older python-six package.  But at least I will be able to migrate to Ubuntu Trusty this way :)

Rest of code looks fine for the python 3/six newbie I am so +1.

Patch-set: 14