Update patch set 2
Patch Set 2: Code-Review-1

(4 comments)

It is possible to avoid the dynamic function calls, but whether the alternative is more/less readable I'm not entirely sure.

The one noticeable benefit is that using a lookup table allows for a JenkinsJobsException to be raised complaining about the "content type" instead of a "KeyError: 'build_pollurl_dummy_content_type'" that would occur if someone accidentally set the content type in the yaml code by the name 'dummy'.

I think that's potentially confusing enough to ask that you at least wrap the call to "function = globals()[function_name]" with a try/except and raise an error message that doesn't require the user to look at the code to understand :)



You can of course improve on the inline comments further as well. I just left the inline ones as least invasive as possible in the patch. Another approach below that also produces the same results.

function:

  def build_one_pollurl_content_type(xml_parent, entries, prefix,
                                     collection_name, element_name):
  
      if entries:
          namespace = 'org.jenkinsci.plugins.urltrigger.content'
          content_type = XML.SubElement(
              xml_parent, '{0}.{1}ContentType'.format(namespace, prefix))
      
          try:
              iterator = iter(entries)
          except TypeError:
              return
          else:
              collection = XML.SubElement(content_type, collection_name)
              for entry in iterator:
                  content_entry = XML.SubElement(
                      collection, '{0}.{1}ContentEntry'.format(namespace, prefix))
                  XML.SubElement(content_entry, element_name).text = entry

In two minds to whether its better to use the try/except approach above or just test for 'hasattr(entries, "__iter__")'

Lookup dict would become:

      valid_content_types = {
          'simple': ['Simple', '', ''],
          'json': ['JSON', 'jsonPaths', 'jsonPath'],
          'text': ['TEXT', 'regExElements', 'regEx'],
          'xml': ['XML', 'xPaths', 'xPath']
      }


And the loop calling function:

          for entry in check_content:
              type_name = entry.keys()[0]
              if type_name not in valid_content_types:
                  raise JenkinsJobsException('pollurl: Invalid Content Type')
  
              content_type = valid_content_types.get(type_name)
              build_one_pollurl_content_type(content_types,
                                             entry[type_name],
                                             *content_type)

Patch-set: 2
Label: Code-Review=-1