Update patch set 2
Patch Set 2: Code-Review-1

Not quite sure what the that this is intended to solve, as it is possible to escape the braces when dealing with shell scripts in job-template definitions.

If you have a script or use braces within job-templates, you need to double them to escape them for the python format function.

i.e.:

 WORKSPACE=${HOME}

becomes

 WORKSPACE=${{HOME}}

Please take a look at http://ci.openstack.org/jenkins-job-builder/configuration.html#job-template and if this doesn't work for you, please elaborate by providing an example test that demonstrates a yaml definition that works with your change but would fail with the existing code even when escaping the braces.

Patch-set: 2
Reviewer: Gerrit User 1054 <1054@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1