Update patch set 6
Patch Set 6:

(1 comment)

The test that is printing that failure message is 'test_non_existing_config_dir'.  JJB throws an error if an invalid directory or file is passed to jenkins-jobs command.  The purpose of the test is to check that an error is thrown.  I think JJB displays the message before throwing the exception.

Patch-set: 6