Update patch set 6
Patch Set 6: Code-Review-1

(4 comments)

Thanks for getting to this!

Just two small things needed, one checking for options.recursive will result in an exception for the delete/delete-all commands and I think the warning message should be moved to be issued in all cases that the input was not a list since file-like objects are valid input to the load_files method as well.

Patch-set: 6
Label: Code-Review=-1