Update patch set 14
Patch Set 14: Code-Review-1

(7 comments)

Looking good, unfortunately in the time that has passed more changes have been added to JJB that require alteration to be python 3 compatible.

Welcome to whack-a-python-3-bug! :)


As a side note I've noticed that the order of a dict's keys becomes truly un-ordered in python 3 and as a result can vary from one run to another, which results in different tests failing on each run since in many cases the XML being generated is dependent on the order of the keys being read from the dicts created from the yaml configuration files.

On the plus side, it shouldn't be too hard to use Ib90a07043112d4739d6529ceddbc9817668bcec0 once it's merged as the basis to override the default mapping to use OrderedDict when reading in the yaml which would guarantee predictability.

Patch-set: 14
Label: Code-Review=-1