Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

My main concern is that there is potential for silently generating incomplete/incorrect jobs leaving the user needing to examine the resulting XML in order to spot problems.

Perhaps making this configurable behaviour, as I would prefer that JJB aborted rather than continued on silently. Definitely think that some tests to act as examples of where this helps and more detail in the commit message or a comment to help with understanding the use case.

Patch-set: 2
Label: Code-Review=-1