Update patch set 3
Patch Set 3: Code-Review+1

Looks good.

FYI, in general I would keep the commit message as describing the purpose of the change rather than using it ask for feedback or trying to track modifications. It's not important, just useful to consider whether the information in the message will actually make sense to someone reading it in 6 months time and it still imparts useful context about the change then.

Other people will differ in their usage, I just prefer that the information there matches and explains the final result rather than how we got there as I would have to refer to gerrit anyway to see the previous incarnation of the change to understand the last line fully.

Patch-set: 3
Label: Code-Review=+1