Update patch set 4
Patch Set 4:

In the example above, the arch: amd64 miss a level of indentation and is not passed to the job. Thus:

      jobs:
         - '{component}-test-{arch}':
           arch: 'amd64'
         - '{component}-test-{arch}'

Is the same as:

 
      jobs:
         - '{component}-test-{arch}'
         - '{component}-test-{arch}'

And the amd64 is not generated.

I fixed the indentation, cherry-picked the other patch and all tests pass as well as the example you posted.


I am rebasing my change on top of the other one.

Patch-set: 4