Update patch set 5
Patch Set 5: Code-Review-1

Extending the test in Id8d7234ce75469cfb933fbcc0d9f5d6c57e26733 shows that these two patches result in slightly unexpected behaviour that's likely to catch out users.

Given that the intention is if you name the defaults in the project specification you can override the used defaults subsequently in the jobs list. If allowing parameters to control which defaults are applied, it would follow through that the user would still expect any overrides in the job list to continue to work.

I suspect that it's to do with the order that the defaults are being applied by each change and what is happening that the directly named defaults and overrides are applied before the templated parameterized named defaults are applied thus meaning anything in the last set wins out.

Not quite sure what needs to be done to have it work as one might expect. I'll try looking at it in more detail to try and see which patch would be the best place to fix this.

Patch-set: 5