Update patch set 2
Patch Set 2: Code-Review-1

(7 comments)

I think you'll have to add in another helper function since the publish-over-ssh functionality can be called as a build-step, build wrapper or publisher step.

It's just the first 1-2 xml tags in each case that differ, then the publish ssh tags are the same which would then allow calling the base_publish_over function.

See the inline comment on the difference between the XML that should be generated. Think it might help to include a test case for the publisher since we're obviously missing one which would have helped show this difference.

Patch-set: 2
Label: Code-Review=-1