Update patch set 7
Patch Set 7: Code-Review-1

(4 comments)

Sorry to nit-pick, but if we're going to fix the commit message lets make sure it is set to something that will be clear when viewing the commit log by itself rather than needing additional context for it to be fully clear.

Unless there is something specific that the commit message should highlight to help explain why the change was being made, it's completely acceptable to simply leave it as just the subject without anything else.

Other than that it looks pretty much ready to merge.

Patch-set: 7
Label: Code-Review=-1