Update patch set 11
Patch Set 11: Code-Review-1

(17 comments)

Sorry it's taken so long to get back to this. Going to try and focus on it to help get it merged soon.

Think it's worth switching the API/CLI too use 'workers' as the argument/option rather than using 'threads' so that any changes to use multiple processes is purely an implementation detail.

Also suggest trying to avoid making any unnecessary changes to the update_job() method as that makes it harder to review and retaining backward compatibility around method names for a few releases.

Few print statements in there that should be either removed, or switched to use logging and potentially moved to be within the parallelize decorator.

Patch-set: 11
Label: Code-Review=-1