Test defaults honors project variables
The defaults can make use of variables passed by the project and I am
taking advantage of that feature in my templates.

With 70ed2236 / https://review.openstack.org/#/c/85336/ there is a
regression which cause the test to fail although it pass with parent
commit d833015.

By moving applyDefaults to be called after deep_format is performed on
the job templates, any substitution contained within defaults is no
longer correctly replaced by project parameters. Must ensure that
deep_format is also called by applyDefaults with the given params in
order to perform the substitution on the applied defaults.

Change-Id: I00fc3700beb56b027b2a870d8239f214f50b0264