Update patch set 3
Patch Set 3:

Do we want to revert the original change responsible for the regression, allow through Antoines subsequent change and combine the work-around here with a re-submission of the original change and see if there are any other options?


Fundamentally though I think that for the originally proposed feature, we have no choice besides performing two levels of substitution. First to apply the name defaults based on a named substitution, and then secondly to apply any project specific params to the named defaults included.

The question is really about whether there is a better way than the current hack.

Not sure if there is, but I think that if this seems a little too hackish, we should remove the regression for now and see what we can do to avoid the regression in re-adding the functionality.

Patch-set: 3
Reviewer: Gerrit User 1054 <1054@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0