Use OrderedDict to preserve key order in created dicts
Adapts https://gist.github.com/enaeseth/844388 which ensures data read
by yaml will have it's order retained in a predictable manner across
multiple python versions and different runs of python 3. This is because
python 3 enables hash randomization by default, which results in the
order of keys when iterating dictionaries varying between python 3.3 and
earlier versions as well as varying slightly over each run.

Additionally it seems more sensible to ensure that the order of
generated XML snippets corresponding to the input yaml files is
consistently in the same order as the entries in the source yaml files.

Change-Id: I6bf6d298a2609cc6ddbbc6b02b7f1a04413a5c89