Update patch set 2
Patch Set 2:

(3 comments)

Tempted to have generateXML() make the call to expandYaml() since it seems unlikely that you could every make a call to generateXML() without making a call to expandYaml() first, so why not have generateXML do it where we are more interested in the XML.

Also wondering if it might be worth storing the yaml and xml output for each job within the same object, and allow requesting the output format.

Not things that need to fixed for this to go in. just food for thought.

Patch-set: 2
Label: Verified=0