Update patch set 2
Patch Set 2:

Heh, so I initially had the generateXml call the expandYaml stuff. I think it would fit in there fine given you can still generate the xml for just one object as you need. However this feels like it is a better separation. Ideally we'd have two classes, 1 which has no xml (ie just yaml) and 1 which only handles the yaml (ie no xml). But this was a little tricky to implement so this is mostly a starting point (hence the commit message).

As for being able to pick this output this may tie xml+yaml generation together. Ideally we should be able to consume just the yaml or otherwise pass that to the xml class to get the Jenkins specific output.

Patch-set: 2
Label: Verified=0