Update patch set 2
Patch Set 2:

Darragh,

This is pretty awesome. I fixed a bunch of python 3 issues but I couldn't figure out how to fix the XML output ordering issues and this seems to fix it!

```
? tox -e py33,py34
...
  py33: commands succeeded
  py34: commands succeeded
  congratulations :)
```

So that's awesome and will be happy to +1 this after you've amended the commit message to Khai's liking!

Patch-set: 2