Update patch set 26
Patch Set 26:

(3 comments)

Main reason for use of 'next(iter(dict.items()))' is that .items(), .keys(), and .values() methods in python 3 return iterator views of dicts which cannot be indexed, while in python 2 they return lists.

Generally have two approaches available, covert the returned iterable to a list and index using:

 list(dict.items())[0]

or wrap the returned iterable with 'iter()' and call next on it to get the first datum:

 next(iter(dict.items()))

I prefer the second since it should avoid creating a copy of a list in python 2 and should simply pass through the iterator on python 3, so I expect it will be generally the most efficient for both.

Patch-set: 26
Label: Verified=0