Update patch set 8
Patch Set 8:

I was just discussing this patch and the reasons behind it (or behind the PYTHONHASHSEED patch I just contributed), and my XMLMatcher from Nova came up.  If we forklifted that code from Nova into the test suite for JJB, would that also solve this problem, without having to enforce an ordering?

Patch-set: 8