Move extended-choice support to parameters module
This patch moves support for the extended choice plugin into the
parameter module so that it can nicely coexist with the other
parameters.  This leaves the existing extended choice plugin code in the
properties module, but removes the documentation from that location and
adds a deprecation warning.  Without this patch you cannot use the
extended choice parameter plugin with any other types of parameters:

Change-Id: Ibace4284fc1bd391948917979a4223bd3007dff4
Closes-Bug: #1361335