Move extended-choice support to parameters module
This patch fixes support for the extended choice plugin in the
parameter module so that it can nicely coexist with the other
parameters.  This leaves the existing extended choice plugin code in the
properties module, but removes the documentation from that location and
adds a deprecation warning.  This patch also provides parity with the
existing functionality under the properties module.

Change-Id: Ibace4284fc1bd391948917979a4223bd3007dff4
Closes-Bug: #1361335