Update patch set 1
Patch Set 1: Code-Review-1

(3 comments)

Think this will work better if you use the dispatch call to use the code in the parameters module instead of retaining it in both places.

Would be good to add a test case, one to replace the docstring example, and ideally second one showing the correct behaviour.

I do wonder if we should add some extra code to fix this fully if we're only providing a warning so that someone could define an extended-choice property along with other params and it will work, but output a warning.

Patch-set: 1
Label: Code-Review=-1